requirements.txt
fastapi==0.97.0
peewee~=3.16.2
python-telegram-bot~=20.3
pandas~=2.0.2
uvicorn==0.22.0
matplotlib~=3.7.1
